requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
cssselect==1.0.3
cycler==0.10.0
feedfinder2==0.0.4
feedparser==5.2.1
future==0.16.0
idna==2.6
jieba3k==0.35.1
kiwisolver==1.0.1
lxml==4.2.1
matplotlib==2.2.2
newspaper3k==0.2.6
nltk==3.2.5
numpy==1.14.2
oauthlib==2.0.7
palettable==3.1.1
pandas==0.22.0
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.2
python-twitter==3.4.1
pytz==2018.4
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
requests-oauthlib==0.8.0
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
tldextract==2.2.0
tqdm==4.23.3
urllib3==1.22
wordcloud==1.4.1
